requirements.txt
Flask==1.1.2
python-telegram-bot==12.8
black==19.10b0
requests==2.24.0
gunicorn==20.0.4